kernel: Allow modules to be integrated into vmlinux
Due to GKI, Android kernels now have hundreds of modules that have complex
dependencies and init timing requirements which are unknown to the kernel.
This makes it very difficult to build-in all modules in a GKI kernel,
because GKI modules are specifically designed to be out-of-tree modules.

With CONFIG_INTEGRATE_MODULES=y, the kernel config now allows options to
remain as =m even with CONFIG_MODULES=n, so that modularized drivers can be
flagged as such and have their initialization deferred to the module init
syscall that userspace performs for the driver's corresponding .ko anyway.

In other words, when userspace tries to do `insmod some_mod.ko`, even if
some_mod.ko is compiled for a different kernel, the insmod attempt will be
routed to initializing the built-in version of some_mod.ko from the module
init syscall. This is done by doing a lookup on "some_mod" to find the
initcall for the module of that specific name. Module parameters are also
handled automatically.

Note that this only works if module names match between what userspace has
and what the kernel has; i.e., if userspace does `insmod some_mod_new.ko`
and an integrated module by the name of "some_mod_new" doesn't exist in the
kernel, then nothing happens.

This setup makes it very easy to build-in all modules on a GKI kernel to
generate a single image containing all drivers, with all of them getting
initialized at the correct time as if they were real modules.

Due to a constraint in the kconfig system, only one config option can have
the `option modules` attribute, so this change therefore breaks the ability
to build actual .ko modules, since CONFIG_MODULES no longer has that
`option modules` attribute. The attribute is needed in order to allow
configs to be set to =m.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: mcdachpappe <noreference@web.de>