arm64: Disable place-relative 32-bit relocations for GCC LTO
There's no simple way to make GCC LTO order initcalls correctly when
pointers to initcalls are stored as place-relative 32-bit relocations,
since this is done with inline assembly. Disable it when GCC LTO is
enabled; the only loss is that initcall pointers take up 64 bits of space
instead of 32 bits.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: mcdachpappe <noreference@web.de>