mali_kbase: Configure for in-tree builds
For in-tree builds, KBUILD_OPTIONS must be set for the build system to
recognize that there are CONFIG_* options local to this module that need to
be expanded. Configure KBUILD_OPTIONS with all of the CONFIG_* options set
in order to fix the in-tree build.

Additionally, don't set any of the KBUILD_CFLAGS warnings, as KBUILD_CFLAGS
changes are ignored for out-of-tree builds but _not_ for in-tree builds. As
a result, this would enable several warnings which break the build, since
these warnings aren't used in mali upstream and thus aren't fixed.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>

Signed-off-by: mcdachpappe <noreference@web.de>