sched: Relax sched class linker order check
When GCC LTO is enabled, the sched classes are indeed emitted in the
correct order but may not necessarily be adjacent to one another in memory.
As a result, the linkage check in sched_init() is spuriously triggered with
GCC LTO.

Relax the check by strictly checking for the order of each sched class as
intended (i.e., by reverse order of priority).

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: mcdachpappe <noreference@web.de>