eh: Increase hardware fifo size to 4096 pages
With real-time scheduling priority for the compression thread, empirical
measurements show that a hardware fifo size of 4096 pages results in the
hardware fifo never getting exhausted under heavy memory pressure. It's not
desirable for the hardware fifo to be exhausted because it means that EH
isn't doing anything. On the other hand, it's not desirable to set the
hardware fifo size to EH's supported maximum of 32768 pages because that
would consume 128 MiB of RAM to reserve for storing the compression buffer.

Using 4096 pages consumes only 16 MiB of RAM for the compression buffer and
makes it possible to keep EH busy nonstop when memory pressure is high,
improving EH utilization and thus zRAM performance.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: mcdachpappe <noreference@web.de>