aoc: Block "debug" service since it also keeps triggering wakeups
Similar to the "logging" AoC (Always-on Coprocessor) service, the "debug"
service is wakeup-capable and triggers wakeups very frequently, thereby
disrupting deep sleep on the AP:

[  143.370789] PM: pm_system_irq_wakeup: 175 triggered 18300000.mbox
...
[  143.429907] aoc 19000000.aoc: Service "debug" has 32 messages to read on wake

This hurts screen-off battery life significantly, which is unacceptable for
a debug mechanism.

Therefore, block the "debug" service to stop it from triggering frequent
wakeups. This improves screen-off battery life quite a bit.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: mcdachpappe <noreference@web.de>