arm64/s2mpu: Support building pkvm-s2mpu-v9 directly into the image
pkvm-s2mpu-v9 is designed to be built as a module and then loaded early
into the hypervisor (nVHE). This assumption is hardcoded into pkvm-s2mpu-v9
in many places, making it impossible to build pkvm-s2mpu-v9 directly into
the kernel image when modules are disabled.

Support building pkvm-s2mpu-v9 directly into the image by building it into
nVHE from the get-go so that it doesn't need to be loaded into nVHE at
runtime, and fix up the various hardcoded modular assumptions so that
pkvm-s2mpu-v9 actually works when it's directly built in.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: mcdachpappe <noreference@web.de>