eh: Overhaul software fifo queue to take unlimited requests
The software fifo queue has an arbitrary limit derived from the number of
request structures pre-allocated in advance. When the queue is full, a
gnarly io_schedule_timeout() is employed to wait until a pre-allocated
request structure is available.

This io_schedule_timeout() is not only slow, but also creates a priority
inversion because the compression thread won't have its priority elevated
to unblock the io_schedule_timeout() waiter. Given that the compression
thread is SCHED_NORMAL, this priority inversion can be quite severe.

Since each zcomp cookie structure contains the page to compress in addition
to a convenient list member which is available to use, ditch the request
pre-allocation entirely and just use zcomp cookies directly. This only
requires ensuring that the zcomp cookie is deleted from the fifo queue
before it's processed by the compression thread, since the cookie's
ownership is transferred back to zcomp after compression is complete. But
until then, the cookie is free for EH to use since zcomp knows that EH
compression is asynchronous.

This neatly eliminates the io_schedule_timeout() usage entirely, greatly
bolstering EH's overall performance.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: mcdachpappe <noreference@web.de>