cal-if: Don't wait for ACPM response when changing DSU or BCI frequency
DSU and BCI frequency are changed often as per the CPU's performance needs;
waiting for an ACPM response to come back for DSU and BCI frequency
transitions therefore adds significant overhead.

Treat DSU and BCI the same way as other subsystems which have rapid DVFS
and don't wait for an ACPM response when their frequencies are changed.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: mcdachpappe <noreference@web.de>