driver core: Adjust probe deferral logic for integrated modules
When INTEGRATE_MODULES is enabled, all module interdependencies are
known to exist (i.e. integrated into the kernel image). Thus, it
always makes sense to return -EPROBE_DEFER when this function is
called, because any missing dependencies will be loaded as an
integrated module.

Doing so avoids volatility when integrated modules are loaded
due to inconsistent probe ordering.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: mcdachpappe <noreference@web.de>