bcmdhd: Fix BT not starting when booting with WiFi turned off
When booting with WiFi toggled off in Android, the WiFi HAL will send a
wlan power off request during boot:
  [    5.152654] wl2wbrc_wlan_off: state -> WLAN_OFF_IN_PROGRESS

BT firmware download will fail when this happens because BT firmware
download requires the WiFi chipset to be powered on. In wbrc_bt_dev_open(),
the BT driver waits for WiFi to be either fully powered on or fully powered
off before proceeding with BT firmware download, which gets stuck
indefinitely because it thinks the WiFi chipset is powering off. This
causes BT to not work until the user manually toggles WiFi on in Android,
at which point the BT firmware download process is unblocked and BT can be
powered on.

This issue occurs because wl_android_wifi_off(), which is the path invoked
by the WiFi HAL in Android to power off WiFi, may not inform the BT driver
that WiFi power-off is finished, despite informing it that WiFi is actively
powering off.

Fix this by ensuring a complementary call to wl2wbrc_wlan_off_finished()
always follows wl2wbrc_wlan_off(), so that the BT driver isn't stuck
indefinitely waiting for WiFi to finish powering off.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: mcdachpappe <noreference@web.de>