vh: metrics: Remove perf metrics bloat
The thermal metrics are utilized for important thermal monitoring by spmic,
in order to filter out spurious temperature readings. Therefore, thermal
metrics are needed; however, perf metrics are not necessary.

Remove the perf metrics while keeping the thermal metrics intact so that it
can be used for crucial spmic functionality.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: mcdachpappe <noreference@web.de>