vh: metrics: Fix thermal uevent when built as an integrated module
The thermal metrics uevent is used by the pixelstats service to observe and
report abnormal thermal readings. It utilizes the pixel_metric module's
uevent interface to receive the abnormal thermal event from the kernel,
which isn't available when thermal metrics is built as an integrated module
(as part of pixel_metrics).

Fix it by creating the module uevent manually when built as an integrated
module.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: mcdachpappe <noreference@web.de>