arm64: dts: zumapro: Don't build zumapro dtb overlays
The SoC dtb overlays for zumapro are for development platforms
(zumapro-emulator.dtbo and zumapro-hybrid.dtbo). Don't build them, and
instead only build the dtbs themselves.

Since the overlays aren't built anymore, we need to specify the symbol
generation flag for dtc since it won't be applied automatically. It's only
automatically applied for dtbs which are known to the build system to be
base dtbs upon which overlays are applied:
  # Set -@ if the target is a base DTB that overlay is applied onto
  DTC_FLAGS += $(if $(filter $(patsubst $(obj)/%,%,$@), $(base-dtb-y)), -@)

But just because we're not building the dtbo doesn't mean we don't use one.
The bootloader still needs dtb symbols to apply the dtbo over the dtb.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: mcdachpappe <noreference@web.de>