tensor_aio: Allow partial updates to saved PMU data
Waiting for a contiguous sample window that meets the minimum sample window
threshold takes an indefinite amount of time, leading to delayed and
incomplete statistics since any shorter sample windows are simply ignored
in the meantime.

Instead of disregarding sample windows that are insufficiently long, let
them be accumulated until they collectively meet the minimum sample window
threshold. This not only improves the accuracy of the CPU's calculated
frequency, but also improves the accuracy of memperfd's view of the system.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: mcdachpappe <noreference@web.de>