tensor_aio: Reduce minimum CPU sample time to track CPU activity faster
When a CPU's gathered statistics don't meet the minimum sample window time,
memperfd flatly ignores that CPU by assuming that it's idle. However, since
the sample window time is relatively long in comparison to a jiffy, this
means that bursty workloads on a CPU can lead to memperfd not scaling up
performance accordingly for that CPU's needs.

Since Tensor AIO's timekeeping is now extremely accurate and precise,
reduce the minimum CPU sample time from 100 us to 3 us. This not only
grants memperfd a more accurate view of CPU activity, but also improves the
accuracy of the measured CPU frequency reported to the scheduler by virtue
of the finer sample granularity making it possible to report a more up to
date frequency measurement when the sample window is very short.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: mcdachpappe <noreference@web.de>