tensor_aio: Avoid racing with CPU hotplug in memperfd
When a CPU is hotplugged, cpu_active_mask is modified without any RCU
synchronization. As a result, the only synchronization for cpu_active_mask
provided by the hotplug code is the CPU hotplug lock.

Take the CPU hotplug lock in memperfd to ensure a stable view of active
CPUs in the system during a memperfd run.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: mcdachpappe <noreference@web.de>