eh: Add IRQ-affined PM QoS constraint while the compression thread runs
EH often finishes processing a compression in fewer than 20 us, at which
point an interrupt is fired to wake the compression thread. The cpuidle
governor is unaware that an interrupt is coming so soon on the CPU with
which the compression IRQ is affined, and therefore hurts both latency and
energy when it chooses a deep idle state with a long residency requirement
and high exit latency.

Optimize EH compression by adding an IRQ-affined PM QoS constraint to keep
the compression IRQ's CPU from entering deep idle states when EH is active.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: mcdachpappe <noreference@web.de>