tensor_aio: Optimize hot for-loops to omit initial condition check
When the compiler cannot prove that a for-loop always iterates at least
once, it emits a branch for the loop's initial condition check.

Since there are several for-loops in hot paths that are guaranteed to have
at least one iteration, the branch for the initial condition check is never
taken and just adds overhead.

Convert eligible hot for-loops into do-while loops to convince the compiler
to omit the initial condition check for them as an optimization.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>

Signed-off-by: mcdachpappe <noreference@web.de>