tensor_aio: Replace idle_cpus cpumask with a simple counter
Using a full cpumask (with the atomic set/clear variants no less) is
overkill for tracking the number of idle CPUs within a domain. Since there
is the stop_machine() scaffolding already in place for the idle-task CPUs
tracker, make use of it in order to optimize out the idle_cpus cpumask with
a simple counter instead, speeding up the very hot idle path.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>

Signed-off-by: mcdachpappe <noreference@web.de>