fs: Manually integrate scope-optimized KernelSU hooks (v1.5)
backslashxx/KernelSU#5

This kernel is in the unique position of technically being a GKI,
that has been de-GKI'ed, and lacks MODULE support. As a result,
newer methods of automatic integration by way of exploiting KPROBES
and/or LKM will not work. Thus, we'll need to adopt the original
approach of hard coding KernelSU hooks as outlined here:

https://kernelsu.org/guide/how-to-integrate-for-non-gki.html

Additionally, refactor these original hooks to be more
appropriately targeted so as to reduce the scope of interception
without interfering with functionality.

Co-Authored-by: backslashxx <118538522+backslashxx@users.noreply.github.com>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>
Signed-off-by: mcdachpappe <noreference@web.de>